Add files via upload
As a part of an academic project, I  calculated the Net working capital of 8 companies over a period of time and compared it with each other and tried to find a solution to optimize it. Developed an understanding of Profit and loss statement, Balance sheet, Net working capital and its impact, Cash flow etc.